Equipments in list and created a callList utils
added the equipments to get ,create and update,

added a separete file to not have 700 lines of code in a file
