+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/repository/EquipmentRepository.kt
@@ -9,4 +9,6 @@ interface EquipmentRepository {
     fun selectEquipment(competitionId: Int, equipmentsId : List<Int>): Boolean
 
     fun verifyEquipmentId(equipmentsId : List<Int>): Boolean
+
+    fun getEquipmentByCompetitionId(competitionId: Int): List<Equipment>
 }
