+++ code/jvm/src/main/kotlin/pt/arbitros/arbnet/services/CallListServiceUtils.kt
@@ -1,0 +1,283 @@
+package pt.arbitros.arbnet.services
+
+import org.springframework.stereotype.Component
+import pt.arbitros.arbnet.domain.CallList
+import pt.arbitros.arbnet.domain.CallListDomain
+import pt.arbitros.arbnet.domain.Competition
+import pt.arbitros.arbnet.domain.ConfirmationStatus
+import pt.arbitros.arbnet.domain.MatchDay
+import pt.arbitros.arbnet.domain.Participant
+import pt.arbitros.arbnet.domain.UtilsDomain
+import pt.arbitros.arbnet.domain.users.User
+import pt.arbitros.arbnet.http.ApiError
+import pt.arbitros.arbnet.http.invalidFieldError
+import pt.arbitros.arbnet.http.model.CallListInputModel
+import pt.arbitros.arbnet.http.model.ParticipantChoice
+import pt.arbitros.arbnet.repository.CallListRepository
+import pt.arbitros.arbnet.repository.CompetitionRepository
+import pt.arbitros.arbnet.repository.FunctionRepository
+import pt.arbitros.arbnet.repository.MatchDayRepository
+import pt.arbitros.arbnet.repository.ParticipantRepository
+import pt.arbitros.arbnet.repository.SessionsRepository
+import pt.arbitros.arbnet.repository.Transaction
+import pt.arbitros.arbnet.repository.UsersRepository
+import pt.arbitros.arbnet.services.Failure
+import java.time.LocalDate
+
+@Component
+class CallListServiceUtils {
+
+    fun validateAndCheckUsers(
+        callList: CallListInputModel,
+        usersRepository: UsersRepository,
+        callListDomain: CallListDomain,
+        utilsDomain: UtilsDomain,
+    ): Either<ApiError, List<User>>? {
+        val validateResult =
+            validateCallList(
+                callList.competitionName,
+                callList.address,
+                callList.phoneNumber,
+                callList.email,
+                callList.association,
+                callList.location,
+                callList.callListType,
+                callListDomain = callListDomain,
+                utilsDomain = utilsDomain,
+            )
+        if (validateResult is Failure) return validateResult
+
+        usersRepository.getUserById(callList.userId)
+            ?: return failure(ApiError.NotFound(
+                "User not found",
+                "No user found with the provided ID",
+            ))
+
+        if(!usersRepository.userHasCouncilRole(callList.userId))
+            return failure(ApiError.InvalidField(
+                "User does not have the required role",
+                "The user must have a council role to create or update a call list"
+            ))
+
+
+        val participants = callList.participants
+        if (participants.isEmpty()) {
+            return success(emptyList())
+        }
+
+        val participantIds = participants.map { it.userId }
+        val foundReferees = usersRepository.getUsersAndCheckIfReferee(participantIds)
+
+        if (foundReferees.size != participants.size) {
+            return failure(ApiError.InvalidField(
+                "Invalid participants",
+                "Some participants are not referees or do not exist in the system",
+            ))
+        }
+        return success(foundReferees)
+    }
+
+    fun createCompetitionAndSessions(
+        callList: CallListInputModel,
+        competitionRepository: CompetitionRepository,
+        matchDayRepository: MatchDayRepository,
+        sessionsRepository: SessionsRepository,
+    ): Pair<Int, Map<LocalDate, Int>> {
+        val competitionId =
+            competitionRepository.createCompetition(
+                callList.competitionName,
+                callList.address,
+                callList.phoneNumber,
+                callList.email,
+                callList.association,
+                callList.location,
+            )
+
+        val matchDayMap =
+            callList
+                .matchDaySessions
+                .associate { md ->
+                    md.matchDay to matchDayRepository.createMatchDay(competitionId, md.matchDay)
+                }
+
+        callList.matchDaySessions.forEach { md ->
+            val mdId = matchDayMap[md.matchDay]!!
+            md.sessions.forEach { tm ->
+                sessionsRepository.createSession(competitionId, mdId, tm)
+            }
+        }
+
+        return competitionId to matchDayMap
+    }
+
+    fun updateCompetitionAndSessions(
+        callList: CallListInputModel,
+        competitionRepository: CompetitionRepository,
+        matchDayRepository: MatchDayRepository,
+        sessionsRepository: SessionsRepository,
+    ): Pair<Int, Map<LocalDate, Int>> {
+        val competitionId =
+            competitionRepository.updateCompetition(
+                callList.callListId!!,  // This function is only called when callList.callListId is not null
+                callList.competitionName,
+                callList.address,
+                callList.phoneNumber,
+                callList.email,
+                callList.association,
+                callList.location,
+            )
+
+        val matchDayMap =
+            callList
+                .matchDaySessions
+                .associate { md ->
+                    val matchDayId =
+                        matchDayRepository.getMatchDayId(competitionId, md.matchDay)
+                            ?: throw IllegalArgumentException("Match day not found") // TODO ERROR HANDLING
+                    md.matchDay to matchDayRepository.updateMatchDay(matchDayId, competitionId, md.matchDay)
+                }
+
+        callList.matchDaySessions.forEach { md ->
+            val mdId = matchDayMap[md.matchDay]!!
+            md.sessions.forEach { tm ->
+                val sessions = sessionsRepository.getSessionByMatchId(mdId)
+                sessions.forEach {
+                    sessionsRepository.updateSession(it.id, competitionId, mdId, tm)
+                }
+            }
+        }
+        return competitionId to matchDayMap
+    }
+
+    fun createCallListOnly(
+        callList: CallListInputModel,
+        callListRepository: CallListRepository,
+        competitionId: Int,
+    ): Int =
+        callListRepository.createCallList(
+            callList.deadline,
+            callList.userId,
+            competitionId,
+            callList.callListType,
+        )
+
+    fun updateCallListOnly(
+        callList: CallListInputModel,
+        callListRepository: CallListRepository,
+        competitionId: Int,
+    ): Int =
+        callListRepository.updateCallList(
+            callList.callListId!!,
+            callList.deadline,
+            callList.callListType,
+            competitionId,
+        )
+
+    fun createParticipantsOnly(
+        participants: List<ParticipantChoice>?,
+        matchDayMap: Map<LocalDate, Int>,
+        callListId: Int,
+        competitionId: Int,
+        functionRepository: FunctionRepository,
+        participantRepository: ParticipantRepository,
+    ): Either<ApiError, Unit> {
+        val participantsToInsert = mutableListOf<Participant>()
+
+        if (participants != null) {
+            for (p in participants) {
+                for ((day, funcName) in p.participantAndRole) {
+                    if (funcName.isBlank()) continue
+
+                    val funcId =
+                        functionRepository.getFunctionIdByName(funcName)
+                            ?: return failure(ApiError.NotFound(
+                                "Function not found",
+                                "No function found with the name '$funcName'",
+                            ))
+
+                    val mdId =
+                        matchDayMap[day]
+                            ?: return failure(ApiError.NotFound(
+                                "Match day not found",
+                                "No match day found for the date '$day'",
+                            ))
+
+                    participantsToInsert +=
+                        Participant(
+                            callListId = callListId,
+                            matchDayId = mdId,
+                            competitionIdMatchDay = competitionId,
+                            userId = p.userId,
+                            functionId = funcId,
+                            confirmationStatus = ConfirmationStatus.WAITING.value,
+                        )
+                }
+            }
+            participantRepository.batchAddParticipants(participantsToInsert)
+        }
+
+        return success(Unit)
+    }
+
+    fun validateCallList(
+        competitionName: String,
+        address: String,
+        phoneNumber: String,
+        email: String,
+        association: String,
+        location: String,
+        callType: String,
+        callListDomain : CallListDomain,
+        utilsDomain: UtilsDomain,
+    ): Either<ApiError, Unit> {
+        if (!callListDomain.validCallListType(callType)) return failure(invalidFieldError("callType"))
+        if (!utilsDomain.validName(competitionName)) return failure(invalidFieldError("competitionName"))
+        if (!utilsDomain.validAddress(address)) return failure(invalidFieldError("address"))
+        if (!utilsDomain.validPhoneNumber(phoneNumber)) return failure(invalidFieldError("phoneNumber"))
+        if (!utilsDomain.validEmail(email)) return failure(invalidFieldError("email"))
+        if (!utilsDomain.validName(association)) return failure(invalidFieldError("association"))
+        if (!utilsDomain.validName(location)) return failure(invalidFieldError("location"))
+
+        return success(Unit)
+    }
+
+    //TODO this does not need to return an Either, it can return a List<MatchDay>, the list can be empty
+    fun getMatchDaysByCompetitionId(
+        tx: Transaction,
+        competitionId: Int,
+    ): Either<ApiError, List<MatchDay>> {
+        val matchDayRepository = tx.matchDayRepository
+
+        val matchDays =
+            matchDayRepository.getMatchDaysByCompetition(competitionId)
+
+        return success(matchDays)
+    }
+
+    fun getParticipantsByCallList(
+        tx: Transaction,
+        callList: CallList,
+    ): Either<ApiError, List<Participant>> {
+        val participantRepository = tx.participantRepository
+
+        val participants = participantRepository.getParticipantsByCallList(callList.id)
+        return success(participants)
+    }
+
+    fun getCompetitionById(
+        tx: Transaction,
+        competitionId: Int,
+    ): Either<ApiError, Competition> {
+        val competitionRepository = tx.competitionRepository
+
+        val competition =
+            competitionRepository.getCompetitionById(competitionId)
+                ?: return failure(ApiError.NotFound(
+                    "Competition not found",
+                    "No competition found with the ID $competitionId"
+                ))
+
+        return success(competition)
+    }
+
+}
